20.4

+++ app/Http/Controllers/Back/ProductCategoriesController.php
@@ -19,6 +19,9 @@ class ProductCategoriesController extends Controller
      */
     public function index()
     {
+        if(!auth()->user()->ability('admin' , 'manage_product_categories' , 'show_product_categories')){
+            return redirect('admin/index');
+        }
         $categories = ProductCategory::withCount('products')
             ->when(\request()->keyword != null , function($query){
                 $query->search(\request()->keyword);
@@ -38,6 +41,10 @@ class ProductCategoriesController extends Controller
      */
     public function create()
     {
+        if(!auth()->user()->ability('admin' , 'create_product_categories')){
+            return redirect('admin/index');
+        }
+
         $main_categories = ProductCategory::whereNull('parent_id')->get(['id' , 'name']);
         return view('back.product_categories.create' , compact('main_categories'));
     }
@@ -50,6 +57,10 @@ class ProductCategoriesController extends Controller
      */
     public function store(ProductCategoryRequest $request)
     {
+        if(!auth()->user()->ability('admin' , 'create_product_categories')){
+            return redirect('admin/index');
+        }
+
         $input['name'] = $request->name;
         $input['status'] = $request->status;
         $input['parent_id'] = $request->parent_id;
@@ -77,6 +88,9 @@ class ProductCategoriesController extends Controller
      */
     public function show($id)
     {
+        if(!auth()->user()->ability('admin' , 'display_product_categories')){
+            return redirect('admin/index');
+        }
         return view('back.product_categories.show');
     }
 
@@ -88,6 +102,9 @@ class ProductCategoriesController extends Controller
      */
     public function edit(ProductCategory $productCategory)
     {
+        if(!auth()->user()->ability('admin' , 'update_product_categories')){
+            return redirect('admin/index');
+        }
         $main_categories = ProductCategory::whereNull('parent_id')->get(['id' , 'name']);
         return view('back.product_categories.edit' , compact('main_categories' , 'productCategory'));
     }
@@ -101,6 +118,10 @@ class ProductCategoriesController extends Controller
      */
     public function update(ProductCategoryRequest $request, ProductCategory $productCategory)
     {
+        if(!auth()->user()->ability('admin' , 'update_product_categories')){
+            return redirect('admin/index');
+        }
+
         $input['name'] = $request->name;
         $input['slug'] = null ;
         $input['status'] = $request->status;
@@ -133,6 +154,9 @@ class ProductCategoriesController extends Controller
      */
     public function destroy(ProductCategory $productCategory)
     {
+        if(!auth()->user()->ability('admin' , 'delete_product_categories')){
+            return redirect('admin/index');
+        }
         if (File::exists('assets/product_categories/'. $productCategory->cover)){
             unlink('assets/product_categories/' . $productCategory->cover);
         }
@@ -145,6 +169,9 @@ class ProductCategoriesController extends Controller
     }
 
     public function remove_image(Request $request){
+        if(!auth()->user()->ability('admin' , 'delete_product_categories')){
+            return redirect('admin/index');
+        }
         $category = ProductCategory::findOrFail($request->product_category_id);
         if (File::exists('assets/product_categories/'. $category->cover)){
             unlink('assets/product_categories/' . $category->cover);
